Move particles to their own package
